Game: Change ::get_round() to return a size_t

+++ engine/include/Game/Game.h
@@ -46,8 +46,7 @@ public:
     int get_dice_total() const;
 
     size_t turn() const { return m_turn; }
-
-    int get_round() const; // turn() // #players
+    size_t get_round() const; // turn() // #players
 
     int largest_army() const;
     int longest_road() const;
